+++ noodles-cram/CHANGELOG.md
@@ -16,6 +16,8 @@
 
   * record: Change alignment start to a `sam::record::Position`.
 
+  * record: Change next mate alignment start to a `sam::record::Position`.
+
   * record/resolve: Pass compression header rather than substitution matrix.
 
     The compression header includes the substitution matrix in the preservation